requirements.txt
datetime
argparse
wandb
h5py
numpy
torchvision
torch
tqdm
matplotlib
astropy
fitsio
sklearn
scikit-image
Pillow
requests
psutil
packaging
pyyaml
galsim
opencv-python